export portamenti in MIDI export as separate notes
If the frequency of a channel is changed while the channel is still
running, generate a note_off() for the old frequency and a note_on()
for the new frequency.

While this is not as perfect as generating real MIDI portamento
events, it is far better than not to export anything at all.