prevent stop of encoding on very short subsongs
When encountering a 5.1 second long subsong the encoding just stopped.

It looked like gbsplay reached the song end, but the pipe buffer was
not filled once, so the ogg encoder got no data at all.  gbsplay did
not close the pipe but seemed to wait for possible input.
Closing gbsplay by redirecting input from /dev/null solved this for me.