plugout_sdl: Allow SDL to choose a sample buffer size
At least on MINGW it appears that SDL chooses a (non-obvious) sample
buffer size based on the underlying audio device and the requested
sample rate. It seems to be a good fit both latency and performance
wise.

While it doesn't eliminate occasional audio choppiness it reduces it
noticeably when compared to a forced sample buffer size.

Audio queue size is also changed to depend on the sample buffer size
chosen by SDL for consistent latency across sample rates.