Reload envelope volume from NRx2 setting on channel retrigger
Currently the last envelope update value is used so a channel remains
silent on retrigger.

When a channel is triggered the volume is supposed to be reset to the
value previously set via the NRx2 register.
To implement the behaviour a property `env_volume` is added to gbhw which
takes the place of the previous `volume`. It is subsequently used to
update sound output, display, etc.

`volume` assumes the value of the actual NRx2 register contents and is
not affected by envelope updates. On channel trigger it is copied to
`env_volume`.