switch to CodeCov.io GitHub Action
The bash uploader recently got compromised and a deprecation plan is
in effect.

Because our build runs on a public build server for apublic
repository, we don't have any secrets or tokens.
Nothing of value could has been lost :)

Switch to the GitHub Action to have an easier update path:
While the plain bash uploader will be removed, the GitHub
Action is planned to switch to some other uploader soon.
If that happens, we will most likely only have to increment
the version number of the dependency.