Add VGM writer plugout
Currently no song metadata (Artist etc.) is written to the VGM file as
the plugout interface does not provide for that yet.

Debatable: a 1-second delay is written at the end of the VGM dump so
sounds have a chance to decay before a player stops output or moves on
to the next file. However the delay leads to a frozen sound when the VGM
dump ends in the middle of playback.