don't accumulate cycle rounding errors when writing MIDI files
We throw away the lower 14 bits of the cycle counter delta to generate
the MIDI timestamp delta.

When exporting every channel as a single MIDI track using channel
muting, the tracks get out of sync after some time.

To fix this, we advance cycle_prev only up to the generated MIDI
timestamp.  The ignored lower bits will then automatically be included
in the next calculation of cycles_delta, so the rounding errors don't
accumulate over time.