fix playback speed in generated MIDI files
Originally this was just a refactoring to put a proper name on the
magic number `124` in the MIDI header.

Understanding the meaning of the value (TIME_DIVISION of 124 with a
default TEMPO of 500_000 ms) pointed out a serious bug: 3906 ms of
audio data were spread to 4032 ms in the MIDI file, resulting in all
generated MIDI files playing ~3% slower than desired.

This has now been fixed by:

- Extract, name and describe the magic constants that are used for the
  timing information in the generated MIDI files.

- Calculate new values for TEMPO and TIME_DIVISION making the MIDI
  speed match the audio data exactly.

- Explicitely set the tempo at the start of the MIDI track because the
  new TEMPO differs from the default value of 500_000 ms.