Undo 1d3483a7c30b016ee4b7fc26f623ea2e24fd50c4
I got confused somewhere.
Channels 1+2 are certainly in the range 64hz-131072hz
For channel 3 the waveform has to repeat once to get the same frequency
as channels 1+2, so GBSOUND.txt seems to be wrong here and it
is actually 32Hz-65536Hz with a non-repeating waveform and up to 1MHz
with a 2-sample repeating waveform.

Also get rid of the A1HZ/A1MIDI constants as we can use C0HZ/C0MIDI in
the formula instead.