Implement custom rand() function
We've had issues with libc rand() implementation differences breaking
the test, so this replaces it with a simple xorshift64 random
implementation, which is simple and should be good enough for this
purpose.

Taken from https://doi.org/10.18637%2Fjss.v008.i14

"Description of a class of simple, extremely fast random number
generators (RNGs) with periods 2k −1 for k = 32, 64, 96, 128, 160, 192.
These RNGs seem to pass tests of randomness very well"