configure: show when check_libs() bails out
It is very irritating to see a string of messages like
```
checking for foo.h:  not found
checking for foo.h in /usr/include/foo:  ok
checking if we need additional libs for foo :  yes
looking for -lfoo:  ok
```
when after all that `config.mk` still contains `use_foo := no`

Explicitely show when an error occurs even if all previous
checks said "ok" and "yes".

This might happen if the library basically can be used, but gcc fails
because of -pedantic or other flags.