configure: ignore -pedantic in check_libs()
libpipewire-0.3 does not compile cleanly, so check_libs() always
returns "library not available" because of the minerrs check, even
when the library is detected properly and compile and link succeed.

This affects our Linux CI pipeline where we set -pedantic
and some other flags.

Relax the checks a bit to fix this.