Make noise channel more faithful to original hardware
Fixes #114.

Switching the LFSR width on the noise channel without retriggering the
channel can cause our LFSR to run out of 1's, rendering the output
silent until the channel is retriggered (using NR44 bit 7).

This is due to the current implementation of the LFSR that differs from
the Gameboy hardware; it moves the taps and provides only a single point
of feedback whereas hardware keeps the taps in place and adds a second
point of feedback near the middle of the LFSR (creating a "split-off"
7-bit-wide LFSR).

cf. https://gbdev.gg8.se/wiki/articles/Gameboy_sound_hardware#Noise_Channel

The schematic reverse engineered by @furrtek mostly supports this
description, except the taps being XNOR'd, not XOR'd:
https://github.com/furrtek/DMG-CPU-Inside/blob/master/Schematics/20_CHANNEL4.png

However, implementing the LFSR strictly according to the schematic lead
to a similar issue; there is reason to believe that the output of the
"AOI_MUX_2" selecting the additional feedback point is actually non-
inverting as opposed to the schematic:

https://github.com/furrtek/DMG-CPU-Inside/issues/77#issuecomment-977359213

Treating it as an AO_MUX_2 indeed yields correct-sounding output.

We can also assume that the flip-flops composing the shift register are
reset to 0 by their /RESET inputs; this also fits the XNOR tap well.

Here is a before-after comparison of the initial LFSR state (wide mode):

  - old -               - new -

 111111111111111       000000000000000
 011111111111111       100000000000000
 001111111111111       110000000000000
[...]
 000000000000011       111111111111100
 000000000000001       111111111111110
 100000000000000       011111111111111
[...]

Due to 0 initialization and inverted tap output, the state is basically
the inverse of the old version, up to the point of a width change.
As a result the noise channel of most tracks (which do not expose the
bug) is now inverted.

Here is a before-after comparison of the LFSR state of previously
incorrect playback, taken from Castlevania 2:

  - old -               - new -

 000001010101000       111110101010111
 000000101010100       111111010101011
 000000010101010       111111101010101
* at this point the game changes LFSR width to 7
 000000001010101       011111110101010
 000000000101010       001111110010101
 000000000010101       000111110001010
 000000000001010       000011110000101
 000000000000101       000001110000010
 000000000000010       000000110000001
 000000000000001       000000010000000
 000000000000000       100000001000000
 000000000000000       110000001100000

And, indeed, the missing parts are now audible and akin to a real
hardware recording.