clear mapped RAM on subsong change
On a subsong change the player has to reset the CPU, registers, RAM
etc.

Until now the mapped external cartridge RAM was not reset, which could
make state from one subsong spill over into the next.  Depending on
the player routine and subsong sequence this could lead to playback
errors.

This fixes issue #124.