plugout_alsa: rewrite handling of write errors
Getting -EPIPE after pausing is no reason to display an error.  If we
don't write data into the ALSA stream while pausing, we have to expect
a buffer underrun.  Just recover silently from that.

This fixes #127 but we might skip some samples on pause/resume in the
ALSA buffer (not sure about this).