don't use i18n in unit tests
We've just added all -I flags to fix the unit tests on brew/MacOS
(common.h includes <libintl.h> and thus needs these flags).

But the -I flags are for the BUILDCC, while the unit tests run with
HOSTCC.  This will probably break when cross-compiling.

We don't really need i18n during unit tests (nothing tests the i18n
part, we test our algorithms instead), so we change the #ifdef around
the i18n path to not trigger when we're running the unit tests.
Now the -I flags are no longer needed for compiling the unit tests.

This partially reverts commit 7e5c6081245.