bash completion: support ~/ in filenames for subsong completion
To get the list of available subsongs, we have to call gbsinfo with
the filename from the commandline currently being completed.

We need to both quote that filename (to prevent nasty surprises with
spacial characters or triggering Command Substitution) and to not
quote it (to trigger Tilde Expansion) – this is not possible.

We choose to quote the filename for security reasions and manually
emulate the most often used case of Tilde Expansion: ~/
All other possible Tilde Expansions won't work.