fix parsing  of "loop" configuration variable
The manpage says that non-zero integers map to range loop mode, but
because we map to an enum, 2 mapped to single loop mode instead.

To keep backward compatibility, actually to what the manpage says and
map 2 to range loop mode.

Range mode now can't be configured any more, but we need a new
configuration option to support all loop modes properly anyway.