bugfix: handle errors from plugout_skip_fn
Some plugouts do file operations when changing subsongs that might
fail.  If for example a file could not be opened, the player routine
ignored the error and die plugout would run into a segfault later when
trying to write data into the unopened filehandle.

For now exit the whole player when an error opening a file occurs.

FIXME: This will immediately exit gbsplay and skip resetting the
terminal.  That's not good, but there is no direct code flow to signal
a clean exit yet, this will need some time to figure out.