REBASE ME: fix expand_filename() detection of overlong strings
Also check the string length when copying single chars.
Until now the "too long" detection only activated during the expansion
of placeholders.

Also simply return 0 for success and 1 for failure.
Until now we wanted to return the possible target string length so
that the caller can check if the string would have been longer than
allowed, but this does not work any more if we break off as soon as
the string gets too long.  We also did not really calculate the
correct possible string length because later in the string there might
have been other placeholders that would increase the string length but
our expansion would never get there after the string was already too
long.
A simple flag is enough to signal success or failure.